feat: unify model listing logic into a single v1_models module
Consolidates LiteLLM and GitHub Copilot model fetching logic into a unified
module that supports standard /v1/models endpoints with automatic SSL
verification skipping for local/private IPs and Copilot token exchange.
