+++ README_zh.md
@@ -104,7 +104,7 @@ pip install -e ".[dev]"
     "my-litellm": {
       "endpoint": "https://api.example.com:4142",
       "api_key_env": "API_KEY_LITELLM",
-      "list_models_cmd": "python -m code_assistant_manager.litellm_models",
+      "list_models_cmd": "python -m code_assistant_manager.v1_models",
       "supported_client": "claude,codex,qwen,copilot",
       "description": "我的 LiteLLM 代理"
     }
